Refactoring taints to reduce sprawl
